Divide client module into project module
