+++ lib/lolp/client.rb
@@ -1,21 +1,9 @@
 require 'lolp/connection'
+require 'lolp/project'
 
 module Lolp
   class Client
     extend Lolp::Connection
-
-    class << self
-      def projects
-        connection.get('projects')
-      end
-
-      def create_project(type, params = {})
-        connection.post('projects', params.merge(haco_type: type))
-      end
-
-      def delete_project(name)
-        connection.delete("projects/#{name}")
-      end
-    end
+    extend Lolp::Project
   end
 end
