--- lib/lolp/project_types.rb
@@ -1,7 +1,0 @@
-module Lolp
-  module ProjectTypes
-    PHP = 'php'.freeze
-    WORDPRESS = 'wordpress'.freeze
-    RAILS = 'rails'.freeze
-  end
-end
